Refactor additional rh data to separate file
